Survival trees for landmark analysis

Fits a "dynTree" model.

dynTree(
  formula,
  data,
  id,
  subset,
  ensemble = TRUE,
  groups = NULL,
  trans = TRUE,
  control = list()
)

Arguments

formula
is a formula object, with the response on the left of a '~' operator, and the terms on the right. The response must be a survival object returned by the 'Surv' function.

data
is an optional data frame in which to interpret the variables occurring in the 'formula'.

id
is an optional vector used to identify the longitudinal observations of subject's id. The length of 'id' should be the same as the total number of observations. If 'id' is missing, each row of `data` represents a distinct observation from a subject and all covariates are treated as a baseline covariate.

subset
is an optional vector specifying a subset of observations to be used in the fitting process.

ensemble
is an optional logical value. If TRUE (default), ensemble methods will be fitted. Otherwise, the survival tree will be fitted.

groups
is an optional vector with the length equal to the number of covariates. This vector indicates the covariates to be splitted together in the emsemble method.

trans
indicates whether to perform time-dependent transformation; only makes sense when all covariates are time-independent.

control
a list of control parameters. See 'details' for important special features of control parameters.

Value

An object of S4 class "dynTree" representig the fit, with the following components:

Details

The argument "control" defaults to a list with the following values:

tau

is the maximum follow-up time; default value is the 90th percentile of the unique observed survival times.

maxNode

is the maximum node number allowed to be in the tree; the default value is 500.

numFold

is the number of folds used in the cross-validation. When numFold > 0, the survival tree will be pruned; when numFold = 0, the unpruned survival tree will be presented. The default value is 10.

h

is the smoothing parameter used in the Kernel; the default value is tau / 20.

minSplitTerm

is the minimum number of baseline observations in each terminal node; the default value is 15.

minSplitNode

is the minimum number of baseline observations in each splitable node; the default value is 30.

disc

is a logical vector specifying whether the covariates in formula are discrete (TRUE) or continuous (FALSE). The length of disc should be the same as the number of covariates in formula. When not specified, the dynTree() function assumes continuous covariates for all.

K

is the number of time points on which the concordance measure is computed. A less refined time grids (smaller K) generally yields faster speed but a very small K is not recommended. The default value is 20.

See also

Examples

data(simDat)

#> Warning: data set ‘simDat’ not found

## Fitting a pruned survival tree
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE)

#> Error in ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE): could not find function "ltree"

## Fitting a unpruned survival tree
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE,
        control = list(numFold = 0))

#> Error in ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE,     control = list(numFold = 0)): could not find function "ltree"

if (FALSE) {
## Fitting the ensemble algorithm (default)
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = TRUE)
}