Printing an dynTree object

The function prints an dynTree object. It is a method for the generic function print of class "dynTree".

# S3 method for dynTree
print(x, digits = 5, tree = NULL, ...)

Arguments

x
an dynTree object.

digits
the number of digits of numbers to print.

tree
an optional integer specifying the \(n^{th}\) tree in the forest to print. The function prints the contents of an lforest object by default, if a tree is not specified.

...
for future development.

Examples

data(simDat)

#> Warning: data set ‘simDat’ not found

## Fitting a pruned survival tree
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE)

#> Error in ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE): could not find function "ltree"

## Fitting a unpruned survival tree
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE,
        control = list(numFold = 0))

#> Error in ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE,     control = list(numFold = 0)): could not find function "ltree"

if (FALSE) {
## Fitting the ensemble algorithm (default)
ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = TRUE)
}